WIP target and timer components + api
Integrates with Ajanottaja's interval endpoints to allow users to start,
stop, and view the active interval.

Implements initial design and functionality of a duration target
component that will allow users to view and modify the target duration
for the current day. Will eventually use the target endpoint for setting
duration in the Ajanottaja API.

The design is still rough and needs polish. The duration form component
uses refs and fancy focus modification to facilitate a better duration
input experience. It also supports pasting.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>