Work in progress improved API and config setup
Try to refactor out repeated setup of useSWR. Repeating the fetch
function for the useSWR body is unecessary and adds more code than is
necessary.

The useSWR documentation specifies that the token used to fetch data
should be part of the argument list to the useSWR function.
Unfortunately this is not possible as the token is fetched using an
async function returned by a different react hook, namely useAuth0.

Rather than trying to move the token creation out into the root of the
app or anything like that we can likely use the hour of day (or however
long the validity of the token is) as a proxy for the token itself. This
should force the swr cache to behave as expected.

A new config module is added which imports the VITE-defined build time
env vars and exports them as handy objects or pure strings. This should
make it easier to see all the build time config as they are defined in
one place.

Ensure we don't render the inner dashboard components before the user is
logged in, and no error has happened.

TODO:

Finish the API abstraction functions for fetching and posting data.

Use the superstruct library to validate responses and coerce values to
the correct types, e.g. coercing ISO8601 Duration strings to the Luxon
library's Duration instances for use in the app, and in the opposite
direction coercing values into strings as expected by the API.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>