Improved calendar suspense fallback and tests
App will no longer show the loading indicator used on the dashboard for
the whole calendar page when loading calendar responses. Instead it uses
a more clever mechanism whereby it generates the dates it should show,
displays the calendar grid, and then loads in the actual results when
they become available. There is a slight delay when switching between
months before the targets and intervals are rendered, but this is mostly
not an issue.

The calendar useSWR hook has been configured to not revalidate the
request when mounting the calendar component as this would trigger new
requests each time the user navigates back and forth between months
renderingg the useSWR caching mechanism moot. This also means that the
user will currently need to refresh the page if she wants new results.

In the future a refresh button of some sort should be added.

Finally, some tests for the date utility functions have been added to
document and test the functions. These are run using the Jest test
runner which uses babel for code transpilation. This is not optimal as
vite itself uses the much faster esbuild, but works well enough for now.

esbuild-jest-transform could be a nice solution in the future when it is
more mature. https://www.npmjs.com/package/esbuild-jest-transform

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>