Week calendar and header redesign
A week calendar has been added with initial layout design and interval
rendering. Currently target durations are not rendered in this view.

The global header has been removed in the calendar page and been
replaced with a menu button that in the future will toggle the global
menu. Instead the calendar page now only shows the calendar navigation
menu which is more important.

The menu calendar is not adapted for mobile yet and will likely require
a pretty different view.

Calendar navigation has been refactored and dates are now properly
normalized so that for instance navigating to the previous month takes
you to the first day of that month.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>