Add interval editor for simple interval editing
A new dialog style interval editor is now present to allow users to edit
their already logged intervals. To support this editing two components,
respectively a date and a time picker component is implemented. These
are both implemented as custom components rather than using the built in
date and time input fields from the browser. This is mostly down to
asthetics and better integration with the design.

To make it easy to refresh the state of the calendar when a user edits
an interval a new SwrMutateProvider react provider is implemented. This
provider allows child components in the react tree below the calendar
page component to access the swr mutate function bound to the current
useSWR state so that state can easily be refreshed.

Duration editing is currently not supported. Neither is creation of new
intervals and durations at arbitrary dates.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>