Register intervals and targets on arbitrary dates
The frontend now exposes a global action menu (the plus button on the
bottom right) that can currently be used to access the interval and
target editors. In this context the editors will allow you to register
completely new targets or intervals on arbitrary dates.

A new custom hook has been added that allows mutating swr hooks using a
regex match. This will provide us the means of updating any calendar
hook currently used so that when registering new targets or dates the
calendar will be updated.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>