Initial landing page with demo components
Implements a landing page featuring the various functionality including
the timer, the target duration picker, and the month calendar view. This
should allow people to grasp the purpose of Ajanottaja more easily.

Upgrades react-router-dom to v6. Also refactores the calendar components
to be more reusable. This allows the landing page to use the month
calendar header and body without the calendar navigation bar or
supplying an auth0 context.

Future work should add the statistics components. The design of the
landing page should also be enhanced.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>