Initial supabase-backed frontend setup
Refactors the Ajanottaja frontend to run on top of a new Supabase
backend. The commit includes a signup and signin page as well as a
reworked dashboard page that enables the same functionality as on the
old Clojure HTTP API backend.

The frontend will use react-query (instead of SWR) as react-query
supports mutation hooks. It uses the standard supabase JavaScript
library to integrate with the Supabase backend API.

For models, validation and types Zod is utilized. This library provides
much the same functionality as superstruct, but has a bigger community
behind it.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>