Initial supabase-backed frontend setup (#12)
* Initial supabase-backed frontend setup

Refactors the Ajanottaja frontend to run on top of a new Supabase
backend. The commit includes a signup and signin page as well as a
reworked dashboard page that enables the same functionality as on the
old Clojure HTTP API backend.

The frontend will use react-query (instead of SWR) as react-query
supports mutation hooks. It uses the standard supabase JavaScript
library to integrate with the Supabase backend API.

For models, validation and types Zod is utilized. This library provides
much the same functionality as superstruct, but has a bigger community
behind it.

* Port all pages and components to supabase

Refactors all components to use the newer Supabase API including
statistics and target and tracks insertion, updates, and deletions.

Updates packages and removes old logic.

* Specify node version using nvmrc dotfile

Cloudflare by default uses an ancient version of Node in its Pages build
command environment. To make it use a later one, and to make it clear
which Node version to use for other devs, a node version can be
specified in a dotfile.

* Always use npm legacy-peer-deps option using dotfile

The build would fail in Cloudflare because it gets a dependency
conflict. Locally this has been avoided using the legacy-peer-deps
option, but there is no way to specify the npm install command in the
Cloudflare pages dashboard. To get around the issue we can use an .npmrc
config file.

* Fix redirect after signup

* Use explicit json casts for timestamptz etc

When querying tables, views, and rpc functions the Ajanottaja Supabase
db will now return timestamptz, tstzrange, and interval values in the
normal SQL format if not instructed otherwise. To return ISO 8601 values
the select statement should now typecast to json explicitly.

* Add email verification page

When a user signs up they are redirected to a page telling them to check
for an email verification mail to complete their account signup.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>