feat: complete design overhaul and port to tailwindcss (#20)
Do a complete design overhaul with improved UX and usability.

Ports project from WindiCSS to TailwindCSS to be more conformant with the current industry standard. TailwindCSS DX has improved since its inception, and in version 4 is more than decent with a plugin for Vite and good support in the community. In sum there is little reason to use WindiCSS outside perhaps the nice dev dashboard they provided.

Changed to a horizontal menu bar on top of the page in the logged in view. On mobile a burger menu allows a slide out menu to appear.

All pages including the home, sign-in, sign-out, dashboard, calendar, and statistics pages have been redesigned. Note that the week calendar and day calendar is currently not quite complete design-wise. I still need to revisit these. All dialogs have been redesigned.

In general the dashboard now makes it easier to track time as it will register 8 hours by default. You can still change the target right from the dashboard. Improved statistics UI is also in place.

In the calendar page it is now easier than ever to add tracked time and targets.

Fixed some general build issues.